switch to history.js instead of bbq.js for HTML5 goodness
This makes the AJAX URL changes occur in the path rather than the fragment portion of the URL. This has the nice side-effect that the server will see the step selector portion of the URL when the browser first accesses the site and we can determine the current step from that. It will still need some work to support non-HTML5 browsers, though.