enable redirect_to_step DSL command in before blocks
This does an actual browser redirect (sets location.pathname to be precise) when the first step accessed encounters the command. This simplifies things a great deal over replacing a semi-loaded step in situ, but it doesn't look very nice. Since this is currently used for handling rare situations of jumping too far into a wizard and it *is* called "redirect_to_step", I'm going to leave it as-is for now.